requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
joblib
typer
